:white_check_mark: Test simple case
