+++ src/main.rs
@@ -1,48 +1,12 @@
-use machine::{Instruction, Machine, Program};
-use regex_syntax::hir::{visit, Hir, HirKind, Literal, Visitor};
+use machine::Machine;
+use regex_syntax::hir::visit;
 use regex_syntax::Parser;
 
+use crate::machine::ProgramFactory;
+
 mod machine;
-
-struct ProgramFactory {
-    program: Program,
-}
-
-impl Default for ProgramFactory {
-    fn default() -> Self {
-        Self {
-            program: Vec::new(),
-        }
-    }
-}
-
-impl Visitor for ProgramFactory {
-    type Err = ();
-    type Output = Vec<Instruction>;
-
-    fn visit_pre(&mut self, hir: &Hir) -> Result<(), Self::Err> {
-        match hir.kind() {
-            HirKind::Concat(_) => {}
-            HirKind::Literal(literal) => match literal {
-                Literal::Unicode(c) => {
-                    self.program.push(Instruction::Char(*c as u8));
-                }
-                Literal::Byte(b) => {
-                    self.program.push(Instruction::Char(*b));
-                }
-            },
-            HirKind::Empty => {
-                self.program.push(Instruction::Match);
-            }
-            _ => todo!(),
-        }
-        Ok(())
-    }
-
-    fn finish(self) -> Result<Self::Output, Self::Err> {
-        Ok(self.program)
-    }
-}
+#[cfg(test)]
+mod tests;
 
 fn main() {
     let hir = Parser::new().parse(r"abc").unwrap();
